requirements.txt
boto3==1.9.177
Flask==1.0.3
gunicorn==19.9.0
lambda-packages==0.20.0
nameparser==1.0.3
phonenumbers==8.10.13
pynamodb==3.3.3
python-dateutil==2.6.0
xmltodict==0.12.0
zappa==0.48.2
